* Melee, ranged and casting haste effects should now stack correctly * Bloodlust increases melee and ranged attack speed * Shamanistic focus will be consumed only by shock spells. Author: Sunda
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@214 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/ascent-world/Player.h
@@ -1245,7 +1245,7 @@ public:
 	void EventAllowTiggerPort(bool enable);
 	float m_rangedattackspeedmod;
 	float m_rangedattackspeedmodammo;
-	int32 m_meleeattackspeedmod;
+	float m_meleeattackspeedmod;
 	uint32 m_modblockabsorbvalue;
 	uint32 m_modblockvaluefromspells;
 	void SendInitialLogonPackets();
